requirements.txt
Flask 
geopy
groq
langchain
langchain-groq
python-dotenv